+++ engine/src/main/scala/cromwell/engine/backend/EngineFunctions.scala
@@ -1,4 +1,4 @@
-package cromwell.engine
+package cromwell.engine.backend
 
 import java.nio.file.FileSystem
