+++ engine/src/main/scala/cromwell/engine/backend/package.scala
@@ -4,6 +4,9 @@ import scala.concurrent.{ExecutionContext, Future}
 import scala.collection.immutable.Seq
 
 package object backend {
+  class WorkflowContext(val root: String)
+  class CallContext(override val root: String, val stdout: String, val stderr: String) extends WorkflowContext(root)
+
   // Ordered by shards, and then ordered by attempts
   type AttemptedCallLogs = Seq[Seq[CallLogs]]
   // Grouped by FQNS
